refactor: (#62) ThemeSelector 및 PHASE_TIME 경로 변경 적용
