+++ src/entities/game/index.ts
@@ -1,4 +1,3 @@
-export * from './model/types';
 export type { ItemIconProps } from './ui/ItemIcon';
 export { ItemIcon } from './ui/ItemIcon';
 export { PlayerAvatar } from './ui/PlayerAvatar';
