+++ src/features/game/index.ts
@@ -4,3 +4,4 @@ export { GameHeader } from './ui/GameHeader';
 export { GameSubmitButton } from './ui/GameSubmitButton';
 export { InventoryPanel } from './ui/InventoryPanel';
 export { PlayerSidebar } from './ui/PlayerSidebar';
+export { ThemeSelector } from './ui/ThemeSelector';
